Minor changes
1. Hide the "latest" section.
2. Add some safeguards in JS.

It has been so much fun to work on this tools database for over 4 years. Now, with the tons of AI tools and services being created every day, I have to think about what to do with this project.
Especially when these AI tools may start out free and then switch to paid plans and/or limit the free features.